+++ src/main/java/vi/talii/service/PlayerService.java
@@ -7,5 +7,5 @@ import vi.talii.model.Player;
 public interface PlayerService {
 
     Player getPlayerById(int id) throws NoSuchPlayerException;
-    void updatePlayersCash(int id, float cash);
+    void updatePlayersCash(int id, double cash);
 }
